customised nn.Sequential
