+++ anvil/checkpoint.py
@@ -5,13 +5,16 @@ Module for loading neural networks and checkpoints - ensuring a copy of model
 is made so that we don't get unexpected results
 
 """
+from pathlib import Path
+from glob import glob
+
+import torch
+
+from reportengine.compat import yaml
 from copy import deepcopy
 
 from reportengine import collect
 
-from anvil.models import normalising_flow
-
-model = normalising_flow#collect("normalising_flow", ("mixture_indices",))
 
 def loaded_checkpoint(checkpoint):
     if checkpoint is None:
@@ -40,3 +43,76 @@ def current_loss(loaded_checkpoint):
     if loaded_checkpoint is None:
         return None
     return loaded_checkpoint["loss"]
+
+
+class InvalidCheckpointError(Exception):
+    pass
+
+
+class InvalidTrainingOutputError(Exception):
+    pass
+
+
+class TrainingRuncardNotFound(InvalidTrainingOutputError):
+    pass
+
+
+class Checkpoint:
+    """Class which saves and loads checkpoints and allows checkpoints to be
+    sorted"""
+
+    def __init__(self, path: str):
+        self.path = Path(path)
+        try:
+            self.epoch = int(self.path.stem.split("_")[-1])  # should be an int
+        except ValueError:
+            raise InvalidCheckpointError(
+                f"{self.path} does not match expected "
+                "name checkpoint: `checkpoint_<epoch>.pt`"
+            )
+
+    def __lt__(self, other):
+        return self.epoch < other.epoch
+
+    def __repr__(self):
+        return str(self.path)
+
+    def load(self):
+        """Return checkpoint dictionary"""
+        return torch.load(self.path)
+
+
+class TrainingOutput:
+    """Class which acts as container for training output, which is a directory
+    containing training configuration, checkpoints and training logs
+    """
+
+    _loaded_config = None
+
+    def __init__(self, path: str):
+        self.path = Path(path)
+        self.config = self.path / "runcard.yml"
+        if not self.config.is_file():
+            raise TrainingRuncardNotFound(
+                f"Invalid training output, no runcard found at: {self.config}"
+            )
+        self.checkpoints = [
+            Checkpoint(cp_path) for cp_path in glob(f"{self.path}/checkpoints/*")
+        ]
+        self.cp_ids = [cp.epoch for cp in self.checkpoints]
+        self.name = self.path.name
+
+    def get_config(self):
+        if self._loaded_config is None:
+            with open(self.config, "r") as f:
+                self._loaded_config = yaml.safe_load(f)
+        return self._loaded_config
+
+    def as_input(self):
+        inp = dict(self.get_config())  # make copy
+        inp["checkpoints"] = self.checkpoints
+        inp["cp_ids"] = self.cp_ids
+        return inp
+
+    def final_checkpoint(self):
+        return max(self.checkpoints)
