+++ anvil/config.py
@@ -8,11 +8,12 @@ import logging
 from reportengine.report import Config
 from reportengine.configparser import ConfigError, element_of, explicit_node
 
-from anvil.core import TrainingOutput
+from anvil.checkpoint import TrainingOutput
 from anvil.train import OPTIMIZER_OPTIONS, reduce_lr_on_plateau
-from anvil.coupling import LAYER_OPTIONS
+from anvil.coupling import _coupling_layers, LAYER_OPTIONS
 from anvil.geometry import Geometry2D
 from anvil.distributions import BASE_OPTIONS, TARGET_OPTIONS
+from anvil.models import normalising_flow
 
 log = logging.getLogger(__name__)
 
@@ -119,13 +120,13 @@ class ConfigParser(Config):
         """Inverse temperature."""
         return beta
 
-    def produce_i_mixture(self, n_mixture):
-        return 0
+    def parse_n_blocks(self, n: int):
+        return n
 
-    def produce_layer_spec(self, layers: list, *, repeats: int):
-        return layers * repeats
+    def produce_block_indices(self, n_blocks):
+        return [{"i_block": i} for i in range(n_blocks)]
 
-    @element_of("layers")
+    @element_of("layer_block")
     def parse_layer(self, layer):
         if layer["layer_type"] not in LAYER_OPTIONS:
             raise ConfigError
@@ -136,6 +137,14 @@ class ConfigParser(Config):
 
     def produce_mixture_indices(self, n_mixture):
         return [{"i_mixture": i} for i in range(n_mixture)]
+
+    @explicit_node
+    def produce_coupling_layers(self):
+        return _coupling_layers
+
+    @explicit_node
+    def produce_model(self):
+        return normalising_flow
 
     def parse_n_batch(self, nb: int):
         """Batch size for training."""
