+++ anvil/core.py
@@ -1,83 +1,120 @@
+r"""
+coupling.py
 """
-core.py
-
-Module containing core objects specific to lattice projects
-"""
-from pathlib import Path
-from glob import glob
-
 import torch
+import torch.nn as nn
 
-from reportengine.compat import yaml
-
-
-class InvalidCheckpointError(Exception):
-    pass
-
-
-class InvalidTrainingOutputError(Exception):
-    pass
-
-
-class TrainingRuncardNotFound(InvalidTrainingOutputError):
-    pass
-
-class Checkpoint:
-    """Class which saves and loads checkpoints and allows checkpoints to be
-    sorted"""
-
-    def __init__(self, path: str):
-        self.path = Path(path)
-        try:
-            self.epoch = int(self.path.stem.split("_")[-1])  # should be an int
-        except ValueError:
-            raise InvalidCheckpointError(
-                f"{self.path} does not match expected "
-                "name checkpoint: `checkpoint_<epoch>.pt`"
-            )
-
-    def __lt__(self, other):
-        return self.epoch < other.epoch
-
-    def __repr__(self):
-        return str(self.path)
-
-    def load(self):
-        """Return checkpoint dictionary"""
-        return torch.load(self.path)
+ACTIVATION_LAYERS = {
+    "relu": nn.ReLU,
+    "leaky_relu": nn.LeakyReLU,
+    "sigmoid": nn.Sigmoid,
+    "tanh": nn.Tanh,
+    None: nn.Identity,
+}
 
 
-class TrainingOutput:
-    """Class which acts as container for training output, which is a directory
-    containing training configuration, checkpoints and training logs
+class Sequential(nn.Sequential):
+    def forward(self, x_input, log_density):
+        for module in self:
+            x_input, log_density = module(x_input, log_density)
+        return x_input, log_density
+
+
+class NeuralNetwork(nn.Module):
+    """Generic class for neural networks used in coupling layers.
+
+    Networks consist of 'blocks' of
+        - Dense (linear) layer
+        - Batch normalisation layer
+        - Activation function
+
+    Parameters
+    ----------
+    size_in: int
+        Number of nodes in the input layer
+    size_out: int
+        Number of nodes in the output layer
+    hidden_shape: list
+        List specifying the number of nodes in the intermediate layers
+    activation: (str, None)
+        Key representing the activation function used for each layer
+        except the final one.
+    final_activation: (str, None)
+        Key representing the activation function used on the final
+        layer.
+    do_batch_norm: bool
+        Flag dictating whether batch normalisation should be performed
+        before the activation function.
+    name: str
+        A label for the neural network, used for diagnostics.
+
+    Methods
+    -------
+    forward:
+        The forward pass of the network, mapping a batch of input vectors
+        with 'size_in' nodes to a batch of output vectors of 'size_out'
+        nodes.
     """
 
-    _loaded_config = None
+    def __init__(
+        self,
+        size_in: int,
+        size_out: int,
+        hidden_shape: list,
+        activation: (str, None),
+        final_activation: (str, None) = None,
+        do_batch_norm: bool = False,
+        label: str = "network",
+    ):
+        super(NeuralNetwork, self).__init__()
+        self.label = label
+        self.size_in = size_in
+        self.size_out = size_out
+        self.hidden_shape = hidden_shape
 
-    def __init__(self, path: str):
-        self.path = Path(path)
-        self.config = self.path / "runcard.yml"
-        if not self.config.is_file():
-            raise TrainingRuncardNotFound(
-                f"Invalid training output, no runcard found at: {self.config}"
-            )
-        self.checkpoints = [
-            Checkpoint(cp_path) for cp_path in glob(f"{self.path}/checkpoints/*")
+        if do_batch_norm:
+            self.batch_norm = nn.BatchNorm1d
+        else:
+            self.batch_norm = nn.Identity
+
+        self.activation_func = ACTIVATION_LAYERS[activation]
+        self.final_activation_func = ACTIVATION_LAYERS[final_activation]
+
+        # nn.Sequential object containing the network layers
+        self.network = self._construct_network()
+
+    def __str__(self):
+        return f"Network: {self.label}\n------------\n{self.network}"
+
+    def _block(self, f_in, f_out, activation_func):
+        """Constructs a single 'dense block' which maps 'f_in' inputs to
+        'f_out' output features. Returns a list with three elements:
+            - Dense (linear) layer
+            - Batch normalisation (or identity if this is switched off)
+            - Activation function
+        """
+        return [
+            nn.Linear(f_in, f_out),
+            self.batch_norm(f_out),
+            activation_func(),
         ]
-        self.cp_ids = [cp.epoch for cp in self.checkpoints]
-        self.name = self.path.name
 
-    def get_config(self):
-        if self._loaded_config is None:
-            with open(self.config, "r") as f:
-                self._loaded_config = yaml.safe_load(f)
-        return self._loaded_config
+    def _construct_network(self):
+        """Constructs the neural network from multiple calls to _block.
+        Returns a torch.nn.Sequential object which has the 'forward' method built in.
+        """
+        layers = self._block(self.size_in, self.hidden_shape[0], self.activation_func)
+        for f_in, f_out in zip(self.hidden_shape[:-1], self.hidden_shape[1:]):
+            layers += self._block(f_in, f_out, self.activation_func)
+        layers += self._block(
+            self.hidden_shape[-1], self.size_out, self.final_activation_func
+        )
+        return nn.Sequential(*layers)
 
-    def as_input(self):
-        inp = dict(self.get_config())  # make copy
-        inp["checkpoints"] = self.checkpoints
-        inp["cp_ids"] = self.cp_ids
-        return inp
-
-    def final_checkpoint(self):
-        return max(self.checkpoints)
+    def forward(self, x: torch.tensor):
+        """Forward pass of the network.
+        
+        Takes a tensor of shape (n_batch, size_in) and returns a new tensor of
+        shape (n_batch, size_out)
+        """
+        return self.network(x)
